move modelsSchema type to shared-api
