+++ apps/api/package.json
@@ -22,7 +22,6 @@
     "@elysiajs/openapi": "^1.4.11",
     "@hebo/database": "workspace:*",
     "@hebo/shared-api": "workspace:*",
-    "@hebo/shared-data": "workspace:*",
     "crypto-random-string": "^5.0.0",
     "elysia": "^1.4.9",
     "pg": "^8.16.3",
